requirements.txt
fastapi
uvicorn[standard]
pypdf
requests
pandas
python-dotenv
python-multipart
